Create project
This project is a self study project. I'm trying to build up a manager
website with authorization and bootstrap, haml, paginate, etc.  But
there aren't any good choice for me so I'm trying to build up this own
set for my future working with rails. It is based on rails 3.2 stable
version, but as 4.0 is coming soon, I think after my study and refresh
this will also based on 4.0 in the future.